+++ django_fast_utils/__init__.py
@@ -1,5 +1,5 @@
 __title__ = 'Django Fast Utils'
-__version__ = '2.0.2'
+__version__ = '2.0.3'
 __author__ = 'Tiago Silva'
 __license__ = 'MIT'