+++ django_fast_utils/settings.py
@@ -1,0 +1,7 @@
+"""
+Settings used for Django Fast Utils
+"""
+
+DJANGO_FAST_UTILS = {
+    'LOGOUT_URL': ['/logout']
+}